Move eval mutation into own file
